quote_display stashed in /scenes for use
Attach to any scene within a canvas layer. On ready, the display will be invisible. If given the function .start("Some Quote - Some Guy") it will parse the string into a formatted richtextlabel that fades on screen. Don't ask how it works, the hamsters running on the series of wheels lead very private lives. For convenience, when just passing .start() without any variables, the colony of hamsters automatically run a default quote (the one alex sent) to start the game. After the fade in/out, the hamsters clean up the screen and return the scene to invisibility, reinitializing themselves for the next .start("Quote - Author") call